Moved tests from test_api to test_movie_queue
